add React-Hook-Form + Zod for validation

+++ src/actions/actions.ts
@@ -4,6 +4,7 @@ import { Pet } from ".prisma/client";
 import prisma from "@/lib/db";
 import { PetEssentials } from "@/lib/types";
 import { sleep } from "@/lib/utils";
+import { petFormSchema, petIdSchema } from "@/lib/validations";
 import { revalidatePath } from "next/cache";
 
 // export async function addPet(formData) {
@@ -69,11 +70,19 @@ import { revalidatePath } from "next/cache";
 //   revalidatePath("/app", "layout");
 // }
 
-export async function addPet(pet: PetEssentials) {
+export async function addPet(pet: unknown) {
   await sleep(1000);
+
+  const validatedPet = petFormSchema.safeParse(pet);
+  if (!validatedPet.success) {
+    return {
+      message: "Invalid pet data",
+    };
+  }
+
   try {
     await prisma.pet.create({
-      data: pet,
+      data: validatedPet.data,
     });
   } catch (error) {
     return {
@@ -84,14 +93,24 @@ export async function addPet(pet: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function editPet(petId: Pet["id"], newPet: PetEssentials) {
+export async function editPet(petId: unknown, newPet: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+  const validatedPet = petFormSchema.safeParse(newPet);
+
+  if (!validatedPetId.success || !validatedPet.success) {
+    return {
+      message: "Invalid pet data",
+    };
+  }
+
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: newPet,
+      data: validatedPet.data,
     });
   } catch (error) {
     return {
@@ -102,12 +121,21 @@ export async function editPet(petId: Pet["id"], newPet: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function checkoutPet(petId: Pet["id"]) {
+export async function checkoutPet(petId: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+
+  if (!validatedPetId.success) {
+    return {
+      message: "Invalid pet data",
+    };
+  }
+
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   } catch (error) {
